refactor: deduplicate Gemini stream parsing in tests
