+++ harness/tests/copy-skills.test.ts
@@ -5,6 +5,7 @@ import path from 'path';
 import { execSync } from 'child_process';
 import { createIsolatedHome, copySkills, cleanupIsolatedHome } from '../lib/agent-shared.ts';
 import { Agents } from '../config.ts';
+import { parseGeminiStreamOutput } from '../agents/gemini-cli-agent.ts';
 function assertSearchResults(output: string) {
     const results = JSON.parse(output);
     assert.ok(Array.isArray(results), 'Output should be a JSON array');
@@ -103,33 +104,7 @@ test('invoking gemini-cli-agent.ts works end-to-end like in eval suite', { skip:
         assert.ok(fs.existsSync(logPath), 'chat_log.txt should exist');
         
         const logContent = fs.readFileSync(logPath, 'utf8');
-        const lines = logContent.split('\n').filter(line => line.trim() !== '');
-        
-        let skillActivated = false;
-        let searchCalled = false;
-        let retrieveCalled = false;
-        
-        for (const line of lines) {
-            try {
-                const event = JSON.parse(line);
-                if (event.type === 'tool_use') {
-                    if (event.tool_name === 'activate_skill' && event.parameters?.name === 'modern-web-use-cases') {
-                        skillActivated = true;
-                    }
-                    if (event.tool_name === 'run_shell_command') {
-                        const command = event.parameters?.command || '';
-                        if (command.includes('--search')) {
-                            searchCalled = true;
-                        }
-                        if (command.includes('--retrieve')) {
-                            retrieveCalled = true;
-                        }
-                    }
-                }
-            } catch (e) {
-                // Ignore parse errors for partial lines if any
-            }
-        }
+        const { skillActivated, searchCalled, retrieveCalled } = parseGeminiStreamOutput(logContent);
         
         console.log(`\n[Validation State]`);
         console.log(`- Skill Activated: ${skillActivated}`);
